added other parts of the form
For now, I defined separate Form-Components for the 4 parts of the requested form:
1. FormGeneral
2. FormRisks
3. FormApplication
4. FormFellowApplicants

If we would decide to display them among each other (as now),
we should probably merge the schemas into a single schema.

Maybe we will show different pages for the different parts(?). That should be decided with Lena.
In that case, we at least want a reusable wrapper-component around JsonForms, that takes the schema as prop.

Once this is done, we should also think about splitting the schemas into smaller units,
to allow reusage of at the moment redundant parts.