fix enableTranslationHelper
The old implementation using next/router didn't respect isReady.
So it didn't work with SSR (as used with `next dev`).

Instead of fixing it with useEffect, I directly implemented the
suggested solution of showing the button whenever the used token is
authorized to submit translation updates.

This obsoletes the redundant `translationHelper=true`